# Improve performance of integer formatting