… src/libraries/System.Private.CoreLib/src/System/Number.Formatting.cs