Follow-up to https://github.com/dotnet/runtime/pull/76519#discussion_r985377405

1. The DivRem(..., 10) for each digit in the number ends up being the most expensive aspect of formatting.  This employs a trick other formatting libraries use, which is to have a table for all the formatted values between 00 and 99, and to then DivRem(..., 100) to cut the number of operations in half, which for longer values is meaningful.
2. Avoids going through the digit counting path when we know at the call site it won't be needed.
3. Employs a branch-free, table-based lookup for CountDigits(ulong) to go with a similar approach added for uint.

The numbers from thesebenchmarks are almost all wins.  The main exception is that the TryFormat, 1 digit, no format specifier case gets a little slower... it's repeatable on my machine, but also sub-ns.

```C#
private ulong[] _uint32Values = new ulong[]
{
    1, 12, 123, 1234, 12345, 123456, 1234567, 12345678, 123456789, 1234567890,
};
private ulong[] _uint64Values = new ulong[]
{
    1, 12, 123, 1234, 12345, 123456, 1234567, 12345678, 123456789, 1234567890,
    12345678901, 123456789012, 1234567890123, 12345678901234, 123456789012345, 1234567890123456, 12345678901234567, 123456789012345678, 1234567890123456789, 12345678901234567890,
};
private char[] _scratch = new char[32];

[Benchmark]
public int UInt32AllLengthsToString()
{
    int sum = 0;
    foreach (ulong value in _uint32Values) sum += value.ToString().Length;
    return sum;
}

[Benchmark]
public bool UInt32AllLengthsTryFormat()
{
    bool success = true;
    foreach (ulong value in _uint32Values) success |= value.TryFormat(_scratch, out _);
    return success;
}

[Benchmark]
public int UInt64AllLengthsToString()
{
    int sum = 0;
    foreach (ulong value in _uint64Values) sum += value.ToString().Length;
    return sum;
}

[Benchmark]
public bool UInt64AllLengthsTryFormat()
{
    bool success = true;
    foreach (ulong value in _uint64Values) success |= value.TryFormat(_scratch, out _);
    return success;
}

[Benchmark]
[Arguments(1)]
[Arguments(12)]
[Arguments(123)]
[Arguments(1234)]
[Arguments(12345)]
[Arguments(123456)]
[Arguments(1234567)]
[Arguments(12345678)]
[Arguments(123456789)]
[Arguments(1234567890)]
public string UInt32ToString(uint value) => value.ToString();

[Benchmark]
[Arguments(1)]
[Arguments(12)]
[Arguments(123)]
[Arguments(1234)]
[Arguments(12345)]
[Arguments(123456)]
[Arguments(1234567)]
[Arguments(12345678)]
[Arguments(123456789)]
[Arguments(1234567890)]
public bool UInt32TryFormat(uint value) => value.TryFormat(_scratch, out _);

[Benchmark]
[Arguments(1, "D")]
[Arguments(255, "D3")]
[Arguments(12345, "D5")]
[Arguments(1234567890, "D10")]
public bool UInt32TryFormatSpecifier(uint value, string format) => value.TryFormat(_scratch, out _, format);

[Benchmark]
[Arguments(1)]
[Arguments(12)]
[Arguments(123)]
[Arguments(1234)]
[Arguments(12345)]
[Arguments(123456)]
[Arguments(1234567)]
[Arguments(12345678)]
[Arguments(123456789)]
[Arguments(1234567890)]
[Arguments(12345678901)]
[Arguments(123456789012)]
[Arguments(1234567890123)]
[Arguments(12345678901234)]
[Arguments(123456789012345)]
[Arguments(1234567890123456)]
[Arguments(12345678901234567)]
[Arguments(123456789012345678)]
[Arguments(1234567890123456789)]
[Arguments(12345678901234567890)]
public string UInt64ToString(ulong value) => value.ToString();

[Benchmark]
[Arguments(1)]
[Arguments(12)]
[Arguments(123)]
[Arguments(1234)]
[Arguments(12345)]
[Arguments(123456)]
[Arguments(1234567)]
[Arguments(12345678)]
[Arguments(123456789)]
[Arguments(1234567890)]
[Arguments(12345678901)]
[Arguments(123456789012)]
[Arguments(1234567890123)]
[Arguments(12345678901234)]
[Arguments(123456789012345)]
[Arguments(1234567890123456)]
[Arguments(12345678901234567)]
[Arguments(123456789012345678)]
[Arguments(1234567890123456789)]
[Arguments(12345678901234567890)]
public bool UInt64TryFormat(ulong value) => value.TryFormat(_scratch, out _);

[Benchmark]
[Arguments(1ul, "D")]
[Arguments(255ul, "D3")]
[Arguments(12345ul, "D5")]
[Arguments(1234567890, "D10")]
[Arguments(12345678901234567890ul, "D20")]
public bool UInt64TryFormatSpecifier(ulong value, string format) => value.TryFormat(_scratch, out _, format);
```

|                    Method |         Toolchain |                value | format |       Mean |     Median | Ratio |
|-------------------------- |------------------ |--------------------- |------- |-----------:|-----------:|------:|
|  UInt32AllLengthsToString | \main\corerun.exe |                    ? |      ? | 127.515 ns | 127.328 ns |  1.00 |
|  UInt32AllLengthsToString |   \pr\corerun.exe |                    ? |      ? | 109.178 ns | 109.004 ns |  0.86 |
|                           |                   |                      |        |            |            |       |
| UInt32AllLengthsTryFormat | \main\corerun.exe |                    ? |      ? |  83.914 ns |  84.299 ns |  1.00 |
| UInt32AllLengthsTryFormat |   \pr\corerun.exe |                    ? |      ? |  57.439 ns |  56.949 ns |  0.68 |
|                           |                   |                      |        |            |            |       |
|  UInt64AllLengthsToString | \main\corerun.exe |                    ? |      ? | 420.978 ns | 415.873 ns |  1.00 |
|  UInt64AllLengthsToString |   \pr\corerun.exe |                    ? |      ? | 307.218 ns | 305.810 ns |  0.71 |
|                           |                   |                      |        |            |            |       |
| UInt64AllLengthsTryFormat | \main\corerun.exe |                    ? |      ? | 295.845 ns | 294.455 ns |  1.00 |
| UInt64AllLengthsTryFormat |   \pr\corerun.exe |                    ? |      ? | 177.322 ns | 176.300 ns |  0.60 |
|                           |                   |                      |        |            |            |       |
|  UInt32TryFormatSpecifier | \main\corerun.exe |                    1 |      D |  12.108 ns |  12.058 ns |  1.00 |
|  UInt32TryFormatSpecifier |   \pr\corerun.exe |                    1 |      D |  11.463 ns |  11.436 ns |  0.95 |
|                           |                   |                      |        |            |            |       |
|  UInt64TryFormatSpecifier | \main\corerun.exe |                    1 |      D |  12.808 ns |  12.757 ns |  1.00 |
|  UInt64TryFormatSpecifier |   \pr\corerun.exe |                    1 |      D |  11.599 ns |  11.601 ns |  0.91 |
|                           |                   |                      |        |            |            |       |
|            UInt32ToString | \main\corerun.exe |                    1 |      ? |   1.019 ns |   1.010 ns |  1.00 |
|            UInt32ToString |   \pr\corerun.exe |                    1 |      ? |   1.026 ns |   1.003 ns |  1.01 |
|                           |                   |                      |        |            |            |       |
|           UInt32TryFormat | \main\corerun.exe |                    1 |      ? |   1.854 ns |   1.840 ns |  1.00 |
|           UInt32TryFormat |   \pr\corerun.exe |                    1 |      ? |   2.548 ns |   2.543 ns |  1.37 |
|                           |                   |                      |        |            |            |       |
|            UInt64ToString | \main\corerun.exe |                    1 |      ? |   1.586 ns |   1.578 ns |  1.00 |
|            UInt64ToString |   \pr\corerun.exe |                    1 |      ? |   1.535 ns |   1.532 ns |  0.97 |
|                           |                   |                      |        |            |            |       |
|           UInt64TryFormat | \main\corerun.exe |                    1 |      ? |   2.457 ns |   2.452 ns |  1.00 |
|           UInt64TryFormat |   \pr\corerun.exe |                    1 |      ? |   3.296 ns |   3.282 ns |  1.34 |
|                           |                   |                      |        |            |            |       |
|            UInt32ToString | \main\corerun.exe |                   12 |      ? |   9.503 ns |   9.473 ns |  1.00 |
|            UInt32ToString |   \pr\corerun.exe |                   12 |      ? |   9.503 ns |   9.493 ns |  1.00 |
|                           |                   |                      |        |            |            |       |
|           UInt32TryFormat | \main\corerun.exe |                   12 |      ? |   2.773 ns |   2.769 ns |  1.00 |
|           UInt32TryFormat |   \pr\corerun.exe |                   12 |      ? |   2.795 ns |   2.780 ns |  1.01 |
|                           |                   |                      |        |            |            |       |
|            UInt64ToString | \main\corerun.exe |                   12 |      ? |  10.107 ns |  10.103 ns |  1.00 |
|            UInt64ToString |   \pr\corerun.exe |                   12 |      ? |   9.406 ns |   9.416 ns |  0.93 |
|                           |                   |                      |        |            |            |       |
|           UInt64TryFormat | \main\corerun.exe |                   12 |      ? |   3.778 ns |   3.771 ns |  1.00 |
|           UInt64TryFormat |   \pr\corerun.exe |                   12 |      ? |   3.798 ns |   3.778 ns |  1.01 |
|                           |                   |                      |        |            |            |       |
|            UInt32ToString | \main\corerun.exe |                  123 |      ? |  10.188 ns |  10.195 ns |  1.00 |
|            UInt32ToString |   \pr\corerun.exe |                  123 |      ? |  10.194 ns |  10.182 ns |  1.00 |
|                           |                   |                      |        |            |            |       |
|           UInt32TryFormat | \main\corerun.exe |                  123 |      ? |   3.459 ns |   3.451 ns |  1.00 |
|           UInt32TryFormat |   \pr\corerun.exe |                  123 |      ? |   3.407 ns |   3.390 ns |  0.99 |
|                           |                   |                      |        |            |            |       |
|            UInt64ToString | \main\corerun.exe |                  123 |      ? |  10.809 ns |  10.777 ns |  1.00 |
|            UInt64ToString |   \pr\corerun.exe |                  123 |      ? |  10.284 ns |  10.265 ns |  0.95 |
|                           |                   |                      |        |            |            |       |
|           UInt64TryFormat | \main\corerun.exe |                  123 |      ? |   4.772 ns |   4.766 ns |  1.00 |
|           UInt64TryFormat |   \pr\corerun.exe |                  123 |      ? |   4.230 ns |   4.219 ns |  0.89 |
|                           |                   |                      |        |            |            |       |
|  UInt64TryFormatSpecifier | \main\corerun.exe |                12345 |     D5 |  16.781 ns |  16.779 ns |  1.00 |
|  UInt64TryFormatSpecifier |   \pr\corerun.exe |                12345 |     D5 |  16.955 ns |  16.897 ns |  1.01 |
|                           |                   |                      |        |            |            |       |
|  UInt32TryFormatSpecifier | \main\corerun.exe |                  255 |     D3 |  14.805 ns |  14.808 ns |  1.00 |
|  UInt32TryFormatSpecifier |   \pr\corerun.exe |                  255 |     D3 |  12.477 ns |  12.435 ns |  0.84 |
|                           |                   |                      |        |            |            |       |
|            UInt32ToString | \main\corerun.exe |                 1234 |      ? |  10.775 ns |  10.771 ns |  1.00 |
|            UInt32ToString |   \pr\corerun.exe |                 1234 |      ? |  10.641 ns |  10.637 ns |  0.99 |
|                           |                   |                      |        |            |            |       |
|           UInt32TryFormat | \main\corerun.exe |                 1234 |      ? |   4.211 ns |   4.187 ns |  1.00 |
|           UInt32TryFormat |   \pr\corerun.exe |                 1234 |      ? |   3.637 ns |   3.636 ns |  0.86 |
|                           |                   |                      |        |            |            |       |
|            UInt64ToString | \main\corerun.exe |                 1234 |      ? |  12.227 ns |  12.224 ns |  1.00 |
|            UInt64ToString |   \pr\corerun.exe |                 1234 |      ? |  10.310 ns |  10.298 ns |  0.84 |
|                           |                   |                      |        |            |            |       |
|           UInt64TryFormat | \main\corerun.exe |                 1234 |      ? |   6.090 ns |   6.068 ns |  1.00 |
|           UInt64TryFormat |   \pr\corerun.exe |                 1234 |      ? |   4.471 ns |   4.469 ns |  0.74 |
|                           |                   |                      |        |            |            |       |
|  UInt32TryFormatSpecifier | \main\corerun.exe |                12345 |     D5 |  16.310 ns |  16.266 ns |  1.00 |
|  UInt32TryFormatSpecifier |   \pr\corerun.exe |                12345 |     D5 |  13.192 ns |  13.180 ns |  0.81 |
|                           |                   |                      |        |            |            |       |
|            UInt32ToString | \main\corerun.exe |                12345 |      ? |  11.502 ns |  11.499 ns |  1.00 |
|            UInt32ToString |   \pr\corerun.exe |                12345 |      ? |  11.027 ns |  10.963 ns |  0.96 |
|                           |                   |                      |        |            |            |       |
|           UInt32TryFormat | \main\corerun.exe |                12345 |      ? |   4.964 ns |   4.958 ns |  1.00 |
|           UInt32TryFormat |   \pr\corerun.exe |                12345 |      ? |   4.295 ns |   4.287 ns |  0.86 |
|                           |                   |                      |        |            |            |       |
|            UInt64ToString | \main\corerun.exe |                12345 |      ? |  13.390 ns |  13.332 ns |  1.00 |
|            UInt64ToString |   \pr\corerun.exe |                12345 |      ? |  11.527 ns |  11.536 ns |  0.86 |
|                           |                   |                      |        |            |            |       |
|           UInt64TryFormat | \main\corerun.exe |                12345 |      ? |   7.249 ns |   7.241 ns |  1.00 |
|           UInt64TryFormat |   \pr\corerun.exe |                12345 |      ? |   5.703 ns |   5.690 ns |  0.79 |
|                           |                   |                      |        |            |            |       |
|            UInt32ToString | \main\corerun.exe |               123456 |      ? |  13.331 ns |  13.307 ns |  1.00 |
|            UInt32ToString |   \pr\corerun.exe |               123456 |      ? |  12.316 ns |  12.260 ns |  0.92 |
|                           |                   |                      |        |            |            |       |
|           UInt32TryFormat | \main\corerun.exe |               123456 |      ? |   5.853 ns |   5.844 ns |  1.00 |
|           UInt32TryFormat |   \pr\corerun.exe |               123456 |      ? |   4.527 ns |   4.523 ns |  0.77 |
|                           |                   |                      |        |            |            |       |
|            UInt64ToString | \main\corerun.exe |               123456 |      ? |  14.744 ns |  14.736 ns |  1.00 |
|            UInt64ToString |   \pr\corerun.exe |               123456 |      ? |  12.577 ns |  12.558 ns |  0.85 |
|                           |                   |                      |        |            |            |       |
|           UInt64TryFormat | \main\corerun.exe |               123456 |      ? |   8.527 ns |   8.477 ns |  1.00 |
|           UInt64TryFormat |   \pr\corerun.exe |               123456 |      ? |   5.943 ns |   5.939 ns |  0.70 |
|                           |                   |                      |        |            |            |       |
|            UInt32ToString | \main\corerun.exe |              1234567 |      ? |  14.198 ns |  14.181 ns |  1.00 |
|            UInt32ToString |   \pr\corerun.exe |              1234567 |      ? |  13.033 ns |  13.019 ns |  0.92 |
|                           |                   |                      |        |            |            |       |
|           UInt32TryFormat | \main\corerun.exe |              1234567 |      ? |   6.850 ns |   6.792 ns |  1.00 |
|           UInt32TryFormat |   \pr\corerun.exe |              1234567 |      ? |   5.151 ns |   5.135 ns |  0.75 |
|                           |                   |                      |        |            |            |       |
|            UInt64ToString | \main\corerun.exe |              1234567 |      ? |  15.767 ns |  15.794 ns |  1.00 |
|            UInt64ToString |   \pr\corerun.exe |              1234567 |      ? |  13.776 ns |  13.722 ns |  0.87 |
|                           |                   |                      |        |            |            |       |
|           UInt64TryFormat | \main\corerun.exe |              1234567 |      ? |   9.584 ns |   9.585 ns |  1.00 |
|           UInt64TryFormat |   \pr\corerun.exe |              1234567 |      ? |   6.925 ns |   6.874 ns |  0.72 |
|                           |                   |                      |        |            |            |       |
|            UInt32ToString | \main\corerun.exe |             12345678 |      ? |  15.101 ns |  15.073 ns |  1.00 |
|            UInt32ToString |   \pr\corerun.exe |             12345678 |      ? |  13.021 ns |  12.951 ns |  0.86 |
|                           |                   |                      |        |            |            |       |
|           UInt32TryFormat | \main\corerun.exe |             12345678 |      ? |   7.677 ns |   7.667 ns |  1.00 |
|           UInt32TryFormat |   \pr\corerun.exe |             12345678 |      ? |   5.568 ns |   5.504 ns |  0.72 |
|                           |                   |                      |        |            |            |       |
|            UInt64ToString | \main\corerun.exe |             12345678 |      ? |  16.790 ns |  16.739 ns |  1.00 |
|            UInt64ToString |   \pr\corerun.exe |             12345678 |      ? |  13.571 ns |  13.555 ns |  0.81 |
|                           |                   |                      |        |            |            |       |
|           UInt64TryFormat | \main\corerun.exe |             12345678 |      ? |  10.555 ns |  10.514 ns |  1.00 |
|           UInt64TryFormat |   \pr\corerun.exe |             12345678 |      ? |   7.175 ns |   7.176 ns |  0.68 |
|                           |                   |                      |        |            |            |       |
|            UInt32ToString | \main\corerun.exe |            123456789 |      ? |  15.594 ns |  15.568 ns |  1.00 |
|            UInt32ToString |   \pr\corerun.exe |            123456789 |      ? |  13.681 ns |  13.660 ns |  0.88 |
|                           |                   |                      |        |            |            |       |
|           UInt32TryFormat | \main\corerun.exe |            123456789 |      ? |   8.559 ns |   8.530 ns |  1.00 |
|           UInt32TryFormat |   \pr\corerun.exe |            123456789 |      ? |   5.962 ns |   5.959 ns |  0.70 |
|                           |                   |                      |        |            |            |       |
|            UInt64ToString | \main\corerun.exe |            123456789 |      ? |  17.546 ns |  17.532 ns |  1.00 |
|            UInt64ToString |   \pr\corerun.exe |            123456789 |      ? |  15.114 ns |  15.159 ns |  0.86 |
…